triggering EntityMetadataExtension afterBeanDiscovery method on Startup integration tests engineering
