+++ column-parent/cassandra-parent/integration-tests/pom.xml
@@ -18,6 +18,11 @@
         </dependency>
         <dependency>
             <groupId>io.quarkiverse.jnosql</groupId>
+            <artifactId>quarkus-jnosql-column-integration-tests</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkiverse.jnosql</groupId>
             <artifactId>quarkus-jnosql-column-cassandra</artifactId>
             <version>${project.version}</version>
         </dependency>
