+++ column-parent/pom.xml
@@ -16,6 +16,7 @@
         <module>runtime</module>
         <module>deployment</module>
         <module>cassandra-parent</module>
+        <module>integration-tests</module>
     </modules>
 
 </project>
