+++ core-parent/pom.xml
@@ -15,6 +15,7 @@
     <modules>
         <module>runtime</module>
         <module>deployment</module>
+        <module>integration-tests</module>
     </modules>
 
 </project>
