+++ core-parent/runtime/src/main/java/io/quarkiverse/jnosql/core/runtime/QuarkusEntityMetadataExtension.java
@@ -1,9 +1,16 @@
 package io.quarkiverse.jnosql.core.runtime;
 
+import jakarta.enterprise.event.Observes;
+import jakarta.enterprise.event.Startup;
 import jakarta.inject.Singleton;
 
 import org.eclipse.jnosql.mapping.reflection.EntityMetadataExtension;
 
 @Singleton
 public class QuarkusEntityMetadataExtension extends EntityMetadataExtension {
+
+    public void onStart(@Observes Startup ev) {
+        this.afterBeanDiscovery(null);
+    }
+
 }
