--- document-parent/arangodb-parent/integration-tests/src/main/java/io/quarkiverse/jnosql/document/it/Person.java
@@ -1,45 +1,0 @@
-package io.quarkiverse.jnosql.document.it;
-
-import java.util.List;
-
-import jakarta.nosql.Column;
-import jakarta.nosql.Entity;
-import jakarta.nosql.Id;
-
-@Entity
-public class Person {
-
-    @Id
-    private String id;
-
-    @Column
-    private String name;
-
-    @Column
-    private List<String> phones;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public List<String> getPhones() {
-        return phones;
-    }
-
-    public void setPhones(List<String> phones) {
-        this.phones = phones;
-    }
-
-}
