+++ document-parent/arangodb-parent/integration-tests/src/test/java/io/quarkiverse/jnosql/document/arangodb/it/JNoSQLResourceIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.jnosql.document.arangodb.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class JNoSQLResourceIT extends JNoSQLResourceTest {
+}
