+++ document-parent/elasticsearch-parent/integration-tests/pom.xml
@@ -18,6 +18,11 @@
         </dependency>
         <dependency>
             <groupId>io.quarkiverse.jnosql</groupId>
+            <artifactId>quarkus-jnosql-document-integration-tests</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkiverse.jnosql</groupId>
             <artifactId>quarkus-jnosql-document-elasticsearch</artifactId>
             <version>${project.version}</version>
         </dependency>
