+++ document-parent/elasticsearch-parent/integration-tests/src/test/java/io/quarkiverse/jnosql/document/elasticsearch/it/JNoSQLResourceIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.jnosql.document.elasticsearch.it;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class JNoSQLResourceIT extends JNoSQLResourceTest {
+}
