--- keyvalue-parent/arangodb-parent/integration-tests/src/test/java/io/quarkiverse/jnosql/keyvalue/dynamodb/it/JNoSQLResourceIT.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.jnosql.keyvalue.dynamodb.it;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class JNoSQLResourceIT extends JNoSQLResourceTest {
-}
