+++ keyvalue-parent/dynamodb-parent/integration-tests/pom.xml
@@ -18,6 +18,11 @@
         </dependency>
         <dependency>
             <groupId>io.quarkiverse.jnosql</groupId>
+            <artifactId>quarkus-jnosql-keyvalue-integration-tests</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.quarkiverse.jnosql</groupId>
             <artifactId>quarkus-jnosql-keyvalue-dynamodb</artifactId>
             <version>${project.version}</version>
         </dependency>
